Refactoring of event router to better organize routes. Moved event model and repository into models dir
