+++ event_router.js
@@ -1,48 +1,55 @@
 const express = require('express')
 const router = express.Router()
 
-const eventRepository = require('./event_repository');
+const eventRepository = require('./models/event_repository');
 
-router.get('/', (req, res) => {
+const getAllEventsRoute = (req, res) => {
     eventRepository.getAllEvents()
         .then((events) => {
             res.status(200).json(events);
         })
-})
+};
 
-router.post('/', (req, res, next) => {
+const createEventRoute = (req, res, next) => {
     eventRepository.addEvent(req.body)
         .then(createdEvent => {
             res.status(201).json(createdEvent)
         })
         .catch(next)
-});
-
-router.get('/:id', (req, res, next) => {
+};
+const getEventByIdRoute = (req, res, next) => {
     eventRepository.getEventById(req.params.id)
         .then(retrievedEvent => {
             res.json(retrievedEvent);
         })
         .catch(next)
-})
+};
 
-router.put('/:id', (req, res, next) => {
+const updateEventRoute = (req, res, next) => {
     eventRepository.updateEvent(req.params.id, req.body)
         .then(updatedEvent => {
             res.json(updatedEvent)
         })
         .catch(next)
-})
+};
 
-router.delete('/:id', (req, res, next) => {
+const deleteEventRoute = (req, res, next) => {
     eventRepository.deleteEvent(req.params.id)
         .then(events => {
             res.status(202).json(events)
         })
         .catch(next)
-})
+};
 
-router.use(function (err, req, res, next) {
+router.route('/')
+    .get(getAllEventsRoute)
+    .post( createEventRoute);
+router.route('/:id')
+    .get(getEventByIdRoute)
+    .put(updateEventRoute)
+    .delete(deleteEventRoute)
+
+router.use((err, req, res, next) => {
     console.log(`Error in method ${req.method}: ${err.message}`)
     res.status(err.status).json({ error: err.message });
 })
